+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli.Tests/Tools/TypeSpec/TspInitToolTest.cs
@@ -4,6 +4,7 @@ using Microsoft.Extensions.Logging;
 using Azure.Sdk.Tools.Cli.Tools.TypeSpec;
 using Moq;
 using Azure.Sdk.Tools.Cli.Helpers;
+using Azure.Sdk.Tools.Cli.Tests.TestHelpers;
 
 namespace Azure.Sdk.Tools.Cli.Tests.Tools
 {
@@ -14,8 +15,9 @@ namespace Azure.Sdk.Tools.Cli.Tests.Tools
         {
             // Arrange
             var npxHelper = new Mock<INpxHelper>().Object;
+            var fileHelper = new Mock<IFileHelper>().Object;
             var logger = new Mock<ILogger<TypeSpecInitTool>>().Object;
-            var tool = new TypeSpecInitTool(npxHelper, CreateTypeSpecHelper(), logger);
+            var tool = new TypeSpecInitTool(npxHelper, CreateTypeSpecHelper(), fileHelper, logger);
 
             // Act
             var command = tool.GetCommandInstances().First();
@@ -31,8 +33,9 @@ namespace Azure.Sdk.Tools.Cli.Tests.Tools
         public async Task Init_WithInvalidTemplate_ShouldReturnError()
         {
             var npxHelper = new Mock<INpxHelper>().Object;
+            var fileHelper = new Mock<IFileHelper>().Object;
             var logger = new Mock<ILogger<TypeSpecInitTool>>().Object;
-            var tool = new TypeSpecInitTool(npxHelper, CreateTypeSpecHelper(), logger);
+            var tool = new TypeSpecInitTool(npxHelper, CreateTypeSpecHelper(), fileHelper, logger);
 
             var result = await tool.InitTypeSpecProjectAsync(outputDirectory: "never-used", template: "invalid-template", serviceNamespace: "MyService", isCli: false);
 
@@ -48,8 +51,9 @@ namespace Azure.Sdk.Tools.Cli.Tests.Tools
         public async Task Init_WithInvalidServiceNamespace_ShouldReturnError()
         {
             var npxHelper = new Mock<INpxHelper>().Object;
+            var fileHelper = new Mock<IFileHelper>().Object;
             var logger = new Mock<ILogger<TypeSpecInitTool>>().Object;
-            var tool = new TypeSpecInitTool(npxHelper, CreateTypeSpecHelper(), logger);
+            var tool = new TypeSpecInitTool(npxHelper, CreateTypeSpecHelper(), fileHelper, logger);
 
             var result = await tool.InitTypeSpecProjectAsync(outputDirectory: "never-used", template: "azure-core", serviceNamespace: "", isCli: false);
 
@@ -64,78 +68,50 @@ namespace Azure.Sdk.Tools.Cli.Tests.Tools
         public async Task Init_WithNonEmptyDirectory_ShouldReturnError()
         {
             var npxHelper = new Mock<INpxHelper>().Object;
+            var fileHelper = new Mock<IFileHelper>().Object;
             var logger = new Mock<ILogger<TypeSpecInitTool>>().Object;
-            var tool = new TypeSpecInitTool(npxHelper, CreateTypeSpecHelper(), logger);
-            var tempDir = Path.Combine(Path.GetTempPath(), $"test-nonexistent-{Guid.NewGuid()}");
-
-            Directory.CreateDirectory(tempDir);
-
-            try
+            var tool = new TypeSpecInitTool(npxHelper, CreateTypeSpecHelper(), fileHelper, logger);
+            using var tempDir = TempDirectory.Create("test-nonexistent");
+            await File.WriteAllTextAsync(Path.Join(tempDir.DirectoryPath, "somefile.txt"), "some file's contents");
+            var result = await tool.InitTypeSpecProjectAsync(outputDirectory: tempDir.DirectoryPath, template: "azure-core", serviceNamespace: "MyService", isCli: false);
+            Assert.Multiple(() =>
             {
-                await File.WriteAllTextAsync(Path.Join(tempDir, "somefile.txt"), "some file's contents");
-
-                var result = await tool.InitTypeSpecProjectAsync(outputDirectory: tempDir, template: "azure-core", serviceNamespace: "MyService", isCli: false);
-
-                Assert.Multiple(() =>
-                {
-                    Assert.That(result.IsSuccessful, Is.False);
-                    Assert.That(result.ResponseError, Does.Contain("Invalid --output-directory"));
-                });
-            }
-            finally
-            {
-                Directory.Delete(tempDir, true);
-            }
+                Assert.That(result.IsSuccessful, Is.False);
+                Assert.That(result.ResponseError, Does.Contain("Invalid --output-directory"));
+            });
         }
 
         [Test]
         public async Task Init_IncorrectGitRepo()
         {
             var npxHelper = new Mock<INpxHelper>().Object;
+            var fileHelper = new Mock<IFileHelper>().Object;
             var logger = new Mock<ILogger<TypeSpecInitTool>>().Object;
-            var tool = new TypeSpecInitTool(npxHelper, CreateTypeSpecHelper(false), logger);
-            var tempDir = Path.Combine(Path.GetTempPath(), $"test-nonexistent-{Guid.NewGuid()}");
-
-            try
+            var tool = new TypeSpecInitTool(npxHelper, CreateTypeSpecHelper(false), fileHelper, logger);
+            using var tempDir = TempDirectory.Create("test-nonexistent");
+            var result = await tool.InitTypeSpecProjectAsync(outputDirectory: tempDir.DirectoryPath, template: "azure-core", serviceNamespace: "MyService", isCli: false);
+            Assert.Multiple(() =>
             {
-                var result = await tool.InitTypeSpecProjectAsync(outputDirectory: tempDir, template: "azure-core", serviceNamespace: "MyService", isCli: false);
-
-                Assert.Multiple(() =>
-                {
-                    Assert.That(result.IsSuccessful, Is.False);
-                    Assert.That(result.ResponseError, Is.EqualTo($"Failed: Invalid --output-directory, must be under the azure-rest-api-specs or azure-rest-api-specs-pr repo"
-));
-                });
-            }
-            finally
-            {
-                Directory.Delete(tempDir, true);
-            }
+                Assert.That(result.IsSuccessful, Is.False);
+                Assert.That(result.ResponseError, Is.EqualTo($"Failed: Invalid --output-directory, must be under the azure-rest-api-specs or azure-rest-api-specs-pr repo"));
+            });
         }
 
         [Test]
         public async Task Init_NotUnderSpecifications()
         {
             var npxHelper = new Mock<INpxHelper>().Object;
+            var fileHelper = new Mock<IFileHelper>().Object;
             var logger = new Mock<ILogger<TypeSpecInitTool>>().Object;
-            var tool = new TypeSpecInitTool(npxHelper, CreateTypeSpecHelper(true), logger);
-            var tempDir = Path.Combine(Path.GetTempPath(), $"test-nonexistent-{Guid.NewGuid()}");
-
-            try
+            var tool = new TypeSpecInitTool(npxHelper, CreateTypeSpecHelper(true), fileHelper, logger);
+            using var tempDir = TempDirectory.Create("test-nonexistent");
+            var result = await tool.InitTypeSpecProjectAsync(outputDirectory: tempDir.DirectoryPath, template: "azure-core", serviceNamespace: "MyService", isCli: false);
+            Assert.Multiple(() =>
             {
-                var result = await tool.InitTypeSpecProjectAsync(outputDirectory: tempDir, template: "azure-core", serviceNamespace: "MyService", isCli: false);
-
-                Assert.Multiple(() =>
-                {
-                    Assert.That(result.IsSuccessful, Is.False);
-                    Assert.That(result.ResponseError, Does.Contain("Invalid --output-directory"));
-                    Assert.That(result.ResponseError, Is.EqualTo($"Failed: Invalid --output-directory, must be under <azure-rest-api-specs or azure-rest-api-specs-pr>{Path.DirectorySeparatorChar}specification"));
-                });
-            }
-            finally
-            {
-                Directory.Delete(tempDir, true);
-            }
+                Assert.That(result.IsSuccessful, Is.False);
+                Assert.That(result.ResponseError, Does.Contain("Invalid --output-directory"));
+                Assert.That(result.ResponseError, Is.EqualTo($"Failed: Invalid --output-directory, must be under <azure-rest-api-specs or azure-rest-api-specs-pr>{Path.DirectorySeparatorChar}specification"));
+            });
         }
 
         private static ITypeSpecHelper CreateTypeSpecHelper(bool isSpecRepo = false)